fix(22): BLK-04 stop lossy round-trip of priceKopecks through whole roubles
On edit, defaultValues used Math.round(plan.priceKopecks / 100) and
submit shipped values.priceRoubles * 100, so any plan whose price was
not a multiple of 100 kopecks (e.g. 250050 → 2500.50 ₽) was silently
snapped to the nearest 100 kopecks on save. Allow fractional roubles
in the schema, drop the round on initialise (kopecks/100 is exact in
JS for typical values), step the input by 0.01, and round explicitly
on submit so the original kopecks value round-trips. Money is still
stored and transported as integer minor units.

+++ apps/admin-web/src/entities/membership/schema.ts
@@ -25,7 +25,10 @@ export type MembershipsListQueryInput = z.infer<typeof membershipsListQuerySchem
 
 export const membershipPlanFormSchema = z.object({
   name: z.string().min(1, 'Введите название').max(120),
-  priceRoubles: z.number().int().min(0), // UI in roubles, converts to kopecks at submit
+  // UI in roubles, converts to kopecks at submit. Fractional roubles are
+  // allowed so existing plans whose priceKopecks is not a multiple of 100
+  // (e.g. 2500.50 ₽) round-trip without lossy snapping (BLK-04).
+  priceRoubles: z.number().min(0),
   durationDays: z.number().int().positive(),
   active: z.boolean().default(true),
 })
